[all] Add project layout
